+++ CHANGES.md
@@ -12,11 +12,13 @@ Improvements 🙌:
  - Add tooltip for room quick actions
  - Pre-share session keys when opening a room or start typing (#2771)
  - Sending is now queuing by room and not uniquely to the session
+ - Improve Snackbar duration (#2929)
 
 Bugfix 🐛:
  - Try to fix crash about UrlPreview (#2640)
  - Be robust if Event.type is missing (#2946)
  - Snappier message send status
+ - Fix MainActivity display (#2927)
 
 Translations 🗣:
  - All string resources and translations have been moved to the application module. Weblate project for the SDK will be removed.